requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
networkx>=3.1
sqlalchemy>=2.0.0
Pillow>=10.0.0
flask>=2.3.0

